(PC-23008)[PRO] refactor: Refactor `ico-plus-circle` icon.
