+++ pro/src/screens/SignupJourneyForm/Activity/ActivityForm.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 
 import FormLayout from 'components/FormLayout'
 import { SelectOption } from 'custom_types/form'
-import { PlusCircleIcon } from 'icons'
+import { ReactComponent as FullMoreIcon } from 'icons/full-more.svg'
 import trashFilledIcon from 'icons/ico-trash-filled.svg'
 import { Button, CheckboxGroup, Select, TextInput } from 'ui-kit'
 import { ButtonVariant } from 'ui-kit/Button/types'
@@ -88,7 +88,7 @@ const ActivityForm = ({ venueTypes }: ActivityFormProps): JSX.Element => {
 
             <Button
               variant={ButtonVariant.TERNARY}
-              Icon={PlusCircleIcon}
+              Icon={FullMoreIcon}
               onClick={() => {
                 arrayHelpers.push('')
               }}
