+++ pro/src/ui-kit/Icons/Icons.tsx
@@ -165,7 +165,6 @@ const notCleandIcons: IconListItem[] = [
   { src: 'icons/other-offer.svg', viewBox: '0 0 30 30' },
   { src: 'icons/party.svg', viewBox: '0 0 48 48' },
   { src: 'icons/ico-passculture.svg', viewBox: '0 0 30 30' },
-  { src: 'icons/ico-plus-circle.svg', viewBox: '0 0 48 48' },
   { src: 'icons/ico-search.svg', viewBox: '0 0 48 48' },
   { src: 'icons/ico-ticket-plus-full.svg', viewBox: '0 0 20 20' },
   { src: 'icons/ico-trash-filled.svg', viewBox: '0 0 48 48' },
